Adds multiple memberships.
Previously, each member could have a single membership plan stored in the member table.  With this revision, there is now a membership table which can contain several rows for each member corresponding to different plans.

The member add and edit functionality was updated.  Functionality to view and add memberships to a given member was also added.

The original filter functionality was too simplistic to function with the new structure.  The filter functionality has been replaced so that filter options now take the form array(array(<filterName>,<args>),array(<filterName>,<args>)).