Modifies key and member modules to show add key form after member signup.
The member module has been modified to include the cid of the new member in the url parameters.  The key module has been given a key_command() function which intercepts the member_add command and adds "tab=keys" to the redirection url, showing the key add form for the newly created member.  Closes #4.