Add new permission member_list to allow/disallow viewing of all members
If member_list is set, users in the corresponding role can see all members in
the member list. If member_list is not set, the user can only see its own entry
in the list.

This can be useful in situations where users should be able to see their own
payment status, but for data security reasons they should not be able to see who
else is a member. With the existing permissions, this case was not possible.