Fixed issue #354: Plan end dates
Added extra check in SQL to not only check if the membership end date is
null but also if the end date is in the future to be considered
'active'.  Also if the end date is in the past they should be considered
not active.  Also if the start date is in the future then they are
considered not yet active.